winscoreai-auto-github/API-Football-auto/requirements.txt
python-dotenv==1.0.1
requests==2.32.3
firebase-admin==6.5.0
pytz==2024.1
